Avoid allocating new clusters only to write zeroes into them
Some clients seem keen to write zeroes, even though the device is
full of zeroes already. Since block allocations currently incur a
flush they are quite expensive. This patch performs a scan of the
buffer and, if it contains only zeroes, skips the block allocation.
Hopefully this speeds things up a little and results in initially
smaller qcow2 files.

Signed-off-by: David Scott <dave@recoil.org>