Avoid flushing after allocating a cluster
In the most complicated case an allocation consists of:

1. enlarge the file to contain 2 new clusters (one for the L2, one for
   the data). The `ftruncate` call will guarantee the file is extended
   with zeroes.
2. writing a pointer in the L1 table to point at the new L2 block
3. writing a pointer in the new L2 block to point at the data cluster

If the ftruncate is lost then we already assume that missing sectors
on the end of the file are full of zeroes.

If the L1 write is lost then the cluster looks unallocated.

If the L2 write is lost then the cluster looks unallocated.

Apart from possibly leaking the allocated clusters, there's no need to
flush when allocating a cluster. Any partial failure (even after a
permutation of writes) will result in the cluster looking unallocated.

Signed-off-by: David Scott <dave@recoil.org>