Add a Discrete Interval Encoding Tree (DIET)
This is based on the functional perl described by Martin Erwig:

  https://web.engr.oregonstate.edu/~erwig/diet/

The tree stores intervals which are merged if they touch or overlap.
This could be used to store the set of blocks which have been
discarded.

The tree is not currently balanced, so membership queries and inserts
could become expensive if the intervals are added in order.

Signed-off-by: David Scott <dave@recoil.org>