Add a `discard` configuration for `create` and `connect`
By default `discard` is a no-op since it is a considered a hint to
the implementation. A "real" implementation will need some scratch
space to store details about discarded blocks. This patch adds an
an optional argument to `create` and `connect` containing this
scratch space.

This is a backwards-compatible API change.

Signed-off-by: David Scott <dave@recoil.org>