compact: create a plan and print it
A plan consists of:

- block copies from the end of the file to the beginning
- flush to make sure these are persisted
- rewrites to update references to the new block locations
- flush to make sure these are persisted
- a final ftruncate to shrink the file

This patch constructs the data we need but refuses to write to the disk.

Note the slight complication that blocks containing references can
themselves move so we need to track which block has moved where.

Note that by splitting the plan into these two phases, we only need to
flush once after each phase for correctness. The operation should be
restartable if it fails half way through.

Signed-off-by: David Scott <dave@recoil.org>