compact: avoid Not_found if the file has no data blocks at all
Previously we would attempt to find the highest cluster in-use for
data and raise Not_found. In this case we consider the last header
block to be the highest in-use cluster. It's not in the Int64Map
because the header blocks are global roots -- no other blocks explicitly
point to them.

Signed-off-by: David Scott <dave@recoil.org>