Add function `get_next_cluster` to compute `next_cluster`
The invariant is that the `next_cluster` should point to the end of
the file. We ensure this is correct while we are modifying the file
ourselves. If we are "checking" a file which someone else is writing
to concurrently we will need to recompute the `next_cluster` before
concluding that a file is really beyond the bounds of the cluster.

Signed-off-by: David Scott <dave@recoil.org>